wip: refactor(scanning): extract service tests into separate file
